This was also our academic project under Unix programming course. We used 2-D array in shell scripting which was a complete new concept of d
This game is a take on the classic game MineSweeper. The game places the
player in minefield where
the objective is to avoid all the mines laid out under the tiles. The
mines are triggered when the
player uncovers them. The mines are never in the same place,this ensures
that every game is a new
experience. This is strategy game where the player has to find the
patterns and use logical
reasoning to get himself out of this mine field. The game operates on a
co-ordinate input system.
The only win scenario int this game is when the player uncovers all the
tiles except the ones
covering the mines and the player will be awarded his freedom from the
minefield.

The game uses shell script and uses many of the useful commands such as
shuf command, the command
used to give a random number in a given range and declare command, the
command used to declare a
2d array in shell.Some interesting features of shell have been used like
changing the color code
to display different colors on the terminal. The goal of the project is
to gain proficiency in
shell scripting,hence the project was coded in shell script